consolidate repetitive pytest session functionality
